Factor out PyScalarUDF and PyAggregateUDF
